Abstract generate json web token into helper function
