+++ controllers/auth.js
@@ -3,7 +3,9 @@ const express = require("express");
 const bcrypt = require("bcrypt");
 const router = express.Router();
 const User = require("../models/user");
-const jsonwebtoken = require("jsonwebtoken");
+
+// Helper Functions
+const {generateJsonToken} = require("./helper");
 
 //ROUTES
 
@@ -16,21 +18,8 @@ router.post("/login", (req, res) => {
         } else {
             // Check if password is correct
             if (bcrypt.compareSync(req.body.password, foundUser.password)) {
-                // Set payload for json web token
-                const user = {
-                    _id: foundUser._id,
-                    email: foundUser.email,
-                    username: foundUser.username,
-                };
-                // Generate json web token and send to front end
-                jsonwebtoken.sign(
-                    { user },
-                    process.env.JWT_SECRET,
-                    { expiresIn: "1d" },
-                    (err, jwtToken) => {
-                        res.json({ jwtToken });
-                    }
-                );
+                // Generate json web token
+                generateJsonToken(foundUser, res);
             } else {
                 res.status(401).json({ error: "wrong password" });
             }
